Extract media source (wip)
